requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.26.46
botocore==1.29.46
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-email-verification==0.2.2
django-extensions==3.2.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.9.1
Faker==14.2.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.1
jupyter-client==7.3.4
jupyter-core==4.11.1
kombu==5.2.4
matplotlib-inline==0.1.3
mysql-connector-python==8.0.31
mysqlclient==2.1.0
nest-asyncio==1.5.5
numpy==1.23.4
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-utils==0.0.2
packaging==21.3
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.2.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
pytzdata==2020.1
pyzmq==23.2.0
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
